Refactor classes into individual files
